+++ packages/arte-odyssey/src/hooks/debounce/index.test.ts
@@ -1,5 +1,5 @@
 import { renderHook } from 'vitest-browser-react';
-import { useDebounce, useDebouncedCallback } from './index';
+import { useDebounce } from './index';
 
 describe('useDebounce', () => {
   it('初期値をそのまま返す', async () => {
@@ -35,42 +35,5 @@ describe('useDebounce', () => {
     await vi.waitFor(() => {
       expect(result.current).toBe('second');
     });
-  });
-});
-
-describe('useDebouncedCallback', () => {
-  it('delay経過後にコールバックが呼ばれる', async () => {
-    const callback = vi.fn();
-    const { result } = await renderHook(() => {
-      const fn = useDebouncedCallback(callback, 50);
-      return fn;
-    });
-
-    expect(typeof result.current).toBe('function');
-    result.current('arg1');
-
-    expect(callback).not.toHaveBeenCalled();
-
-    await vi.waitFor(() => {
-      expect(callback).toHaveBeenCalledOnce();
-    });
-    expect(callback).toHaveBeenCalledWith('arg1');
-  });
-
-  it('delay内に複数回呼ばれたら最後の呼び出しだけ実行される', async () => {
-    const callback = vi.fn();
-    const { result } = await renderHook(() => {
-      const fn = useDebouncedCallback(callback, 50);
-      return fn;
-    });
-
-    result.current('first');
-    result.current('second');
-    result.current('third');
-
-    await vi.waitFor(() => {
-      expect(callback).toHaveBeenCalledOnce();
-    });
-    expect(callback).toHaveBeenCalledWith('third');
   });
 });
